Description:fix codex bug
Match-id-66938b1d7d9023955a218e7dbe6748ee34c32999

+++ common/include/input_hub.cpp
@@ -994,11 +994,11 @@ void InputHub::HandleTouchScreenEvent(struct input_event readBuffer[], const siz
             struct input_event &iev = readBuffer[j];
             if (iev.code == ABS_MT_POSITION_X || iev.code == ABS_X) {
                 absInfo.absX = iev.value;
-                absInfo.absXIndex = j;
+                absInfo.absXIndex = (int32_t)j;
             }
             if (iev.code == ABS_MT_POSITION_Y || iev.code == ABS_Y) {
                 absInfo.absY = iev.value;
-                absInfo.absYIndex = j;
+                absInfo.absYIndex = (int32_t)j;
             }
         }
